microphone_dma/web/requirements.txt
flask==2.2.5
werkzeug==2.2.3
pyserial==3.5
